# application/api-spec/src/main/resources/openapi/hotel-reservation.yaml
openapi: 3.0.3
info:
  title: Hotel Reservation API
  version: 1.0.0
  description: |
    Hotel Reservation backend API for room availability search, reservation management,
    and staff stay operations.
servers:
  - url: http://localhost:8080
paths:
...
  /staff/reservations:
    post:
      tags:
        - Staff Reservations
      operationId: createStaffReservation
      summary: Create staff reservation
      description: Creates an internal staff/admin reservation. Staff may provide an existing guestId or guest contact data for creating/finding a Guest profile.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffCreateReservationRequest'
            examples:
              forExistingGuest:
                value:
                  hotelId: 1001
                  roomTypeId: 1102
                  guestId: 11
                  checkIn: '2026-05-10'
                  checkOut: '2026-05-12'
                  adults: 2
      responses:
        '201':
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateReservationResponse'
        '400':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /staff/reservations/{reservationId}/check-in:
    post:
      tags:
        - Staff Reservations
      operationId: checkInReservation
      summary: Check in reservation
      security:
        - bearerAuth: []
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reservation checked in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          description: Business rule violation
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Reservation or room not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /staff/reservations/{reservationId}/check-out:
    post:
      tags:
        - Staff Reservations
      operationId: checkOutReservation
      summary: Check out reservation
      security:
        - bearerAuth: []
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reservation checked out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          description: Business rule violation
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Reservation or room not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /staff/reservations/{reservationId}/no-show:
    post:
      tags:
        - Staff Reservations
      operationId: markNoShowReservation
      summary: Mark reservation as no-show
      security:
        - bearerAuth: []
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reservation marked as no-show
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          description: Business rule violation
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Reservation not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /staff/rooms/{roomId}/status:
    patch:
      tags:
        - Staff Rooms
      operationId: updateRoomStatus
      summary: Update room operational status
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomStatusRequest'
      responses:
        '200':
          description: Room status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomOperationResponse'
        '400':
          description: Business rule violation
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Room not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BookingPet:
      type: object
      required:
        - type
        - size
        - weightKg
      properties:
        type:
          type: string
          enum: [DOG, CAT, OTHER]
        size:
          type: string
          enum: [SMALL, MEDIUM, LARGE]
        weightKg:
          type: number
          minimum: 0
...
    StaffCreateReservationRequest:
      type: object
      required:
        - hotelId
        - roomTypeId
        - checkIn
        - checkOut
        - adults
      properties:
        hotelId:
          type: integer
          format: int64
        roomTypeId:
          type: integer
          format: int64
        guestId:
          type: integer
          format: int64
          description: Existing Guest profile id. Staff/admin only.
        firstName:
          type: string
          description: Required with lastName, email, and phone when guestId is not provided.
        lastName:
          type: string
          description: Required with firstName, email, and phone when guestId is not provided.
        email:
          type: string
          format: email
          description: Required with firstName, lastName, and phone when guestId is not provided.
        phone:
          type: string
          description: Required with firstName, lastName, and email when guestId is not provided.
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        adults:
          type: integer
          minimum: 1
        childrenAges:
          type: array
          description: Ages of all accompanying minors, including infants. The backend classifies each age using the selected hotel's infantMaxAge, childMaxAge, and adultEquivalentAge; infants count against maxInfants, children count against maxChildren, and adult-equivalent minors count against maxAdults/maxTotalGuests.
          items:
            type: integer
            minimum: 0
        pets:
          type: array
          items:
            $ref: '#/components/schemas/BookingPet'
        serviceOfferings:
          type: array
          items:
            $ref: '#/components/schemas/ReservationServiceSelection'
    ReservationServiceSelection:
      type: object
      required:
        - serviceOfferingId
        - quantity
      properties:
        serviceOfferingId:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 1
    CreateReservationResponse:
      allOf:
        - $ref: '#/components/schemas/ReservationResponse'
    ReservationResponse:
      type: object
      required:
        - reservationId
        - hotelId
        - guestId
        - roomTypeId
        - checkIn
        - checkOut
        - adults
        - childrenAges
        - pets
        - status
        - createdAt
        - createdBy
      properties:
        reservationId:
          type: string
        hotelId:
          type: integer
          format: int64
        guestId:
          type: integer
          format: int64
        roomId:
          type: integer
          format: int64
          nullable: true
        roomTypeId:
          type: integer
          format: int64
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        adults:
          type: integer
        childrenAges:
          type: array
          description: Ages of all accompanying minors exactly as submitted, including infants. Infant/child classification is derived from the hotel's age policy.
          items:
            type: integer
        pets:
          type: array
          items:
            $ref: '#/components/schemas/BookingPet'
        contactEmail:
          type: string
          format: email
          nullable: true
        contactPhone:
          type: string
          nullable: true
          maxLength: 32
        specialRequests:
          type: string
          nullable: true
          maxLength: 500
        basePriceAmount:
          type: number
        basePriceCurrency:
          type: string
        servicesPriceAmount:
          type: number
        servicesPriceCurrency:
          type: string
        discountAmount:
          type: number
        discountCurrency:
          type: string
        finalPriceAmount:
          type: number
        finalPriceCurrency:
          type: string
        serviceItems:
          type: array
          items:
            $ref: '#/components/schemas/ReservationServiceItemResponse'
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        createdBy:
          type: string
    ReservationServiceItemResponse:
      type: object
      required:
        - serviceOfferingId
        - serviceName
        - priceAmount
        - priceCurrency
        - quantity
        - totalPriceAmount
        - totalPriceCurrency
      properties:
        serviceOfferingId:
          type: integer
          format: int64
        serviceName:
          type: string
        priceAmount:
          type: number
        priceCurrency:
          type: string
        quantity:
          type: integer
        totalPriceAmount:
          type: number
        totalPriceCurrency:
          type: string
    UpdateRoomStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [AVAILABLE, CLEANING, MAINTENANCE, OUT_OF_SERVICE]
    RoomOperationResponse:
      type: object
      required:
        - roomId
        - hotelId
        - roomNumber
        - roomTypeId
        - capacity
        - status
      properties:
        roomId:
          type: integer
          format: int64
        hotelId:
          type: integer
          format: int64
        roomNumber:
          type: string
        roomTypeId:
          type: integer
          format: int64
        capacity:
          type: integer
        status:
          type: string
          enum: [AVAILABLE, OCCUPIED, CLEANING, MAINTENANCE, OUT_OF_SERVICE]
...
    ErrorResponse:
      type: object
      description: RFC 7807 problem details response returned by the REST layer.
      properties:
        type:
          type: string
          example: about:blank
        title:
          type: string
          example: Validation failed
        status:
          type: integer
          format: int32
          example: 400
        detail:
          type: string
          example: checkOut must be after checkIn
        instance:
          type: string
          example: /reservations